CS50 Scrabble
For this exercise in the Harvard University's, CS50, EdX class, we were instructed to create a game of Scrabble.  The program had to receive words from two players, as well as calculate the score of each word based on a predetermined array of numbers.  Each number in the array must give a value to each corresponding letter to properly calculate the score.